Move tests and remove outdated code
