+++ packages/ref-core/tests/unit/test_providers.py
@@ -5,9 +5,9 @@ from pathlib import Path
 import pytest
 
 import cmip_ref_core.providers
-from cmip_ref_core.exceptions import InvalidMetricException
+from cmip_ref_core.exceptions import InvalidMetricException, InvalidProviderException
 from cmip_ref_core.metrics import CommandLineMetric, Metric
-from cmip_ref_core.providers import CondaMetricsProvider, MetricsProvider
+from cmip_ref_core.providers import CondaMetricsProvider, MetricsProvider, import_provider
 
 
 class TestMetricsProvider:
@@ -46,6 +46,37 @@ class TestMetricsProvider:
 
         result = provider.get("mock")
         assert isinstance(result, Metric)
+
+
+@pytest.mark.parametrize("fqn", ["cmip_ref_metrics_esmvaltool.provider", "cmip_ref_metrics_esmvaltool"])
+def test_import_provider(fqn):
+    provider = import_provider(fqn)
+
+    assert provider.name == "ESMValTool"
+    assert provider.slug == "esmvaltool"
+    assert isinstance(provider, MetricsProvider)
+
+
+def test_import_provider_missing():
+    fqn = "cmip_ref"
+    match = f"Invalid provider: '{fqn}'\n Provider 'provider' not found in cmip_ref"
+    with pytest.raises(InvalidProviderException, match=match):
+        import_provider(fqn)
+
+    fqn = "cmip_ref.datasets.WrongProvider"
+    match = f"Invalid provider: '{fqn}'\n Provider 'WrongProvider' not found in cmip_ref.datasets"
+    with pytest.raises(InvalidProviderException, match=match):
+        import_provider(fqn)
+
+    fqn = "missing.local.WrongProvider"
+    match = f"Invalid provider: '{fqn}'\n Module 'missing.local' not found"
+    with pytest.raises(InvalidProviderException, match=match):
+        import_provider(fqn)
+
+    fqn = "cmip_ref.constants.config_filename"
+    match = f"Invalid provider: '{fqn}'\n Expected MetricsProvider, got <class 'str'>"
+    with pytest.raises(InvalidProviderException, match=match):
+        import_provider(fqn)
 
 
 @pytest.mark.parametrize(
