+++ packages/ref/tests/unit/test_provider_registry.py
@@ -1,39 +1,4 @@
-import pytest
-
-from cmip_ref.provider_registry import ProviderRegistry, import_provider
-from cmip_ref_core.exceptions import InvalidProviderException
-from cmip_ref_core.providers import MetricsProvider
-
-
-@pytest.mark.parametrize("fqn", ["cmip_ref_metrics_esmvaltool.provider", "cmip_ref_metrics_esmvaltool"])
-def test_import_provider(fqn):
-    provider = import_provider(fqn)
-
-    assert provider.name == "ESMValTool"
-    assert provider.slug == "esmvaltool"
-    assert isinstance(provider, MetricsProvider)
-
-
-def test_import_provider_missing():
-    fqn = "cmip_ref"
-    match = f"Invalid provider: '{fqn}'\n Provider 'provider' not found in cmip_ref"
-    with pytest.raises(InvalidProviderException, match=match):
-        import_provider(fqn)
-
-    fqn = "cmip_ref.datasets.WrongProvider"
-    match = f"Invalid provider: '{fqn}'\n Provider 'WrongProvider' not found in cmip_ref.datasets"
-    with pytest.raises(InvalidProviderException, match=match):
-        import_provider(fqn)
-
-    fqn = "missing.local.WrongProvider"
-    match = f"Invalid provider: '{fqn}'\n Module 'missing.local' not found"
-    with pytest.raises(InvalidProviderException, match=match):
-        import_provider(fqn)
-
-    fqn = "cmip_ref.constants.config_filename"
-    match = f"Invalid provider: '{fqn}'\n Expected MetricsProvider, got <class 'str'>"
-    with pytest.raises(InvalidProviderException, match=match):
-        import_provider(fqn)
+from cmip_ref.provider_registry import ProviderRegistry
 
 
 class TestProviderRegistry:
